+++ packages/devextreme/testing/testcafe/helpers/createWidget.ts
@@ -1,59 +1,7 @@
 import { ClientFunction } from 'testcafe';
 import type { Properties as DataGridProperties } from '../../../js/ui/data_grid';
 import type { Properties as FilterBuilderProperties } from '../../../js/ui/filter_builder';
-
-export type WidgetName =
-'dxAccordion'
-| 'dxAutocomplete'
-| 'dxGallery'
-| 'dxButtonGroup'
-| 'dxCalendar'
-| 'dxCalendarView'
-| 'dxCheckBox'
-| 'dxColorBox'
-| 'dxDropDownButton'
-| 'dxDraggable'
-| 'dxTabPanel'
-| 'dxForm'
-| 'dxFilterBuilder'
-| 'dxSelectBox'
-| 'dxScrollable'
-| 'dxScrollView'
-| 'dxMultiView'
-| 'dxPivotGrid'
-| 'dxPivotGridFieldChooser'
-| 'dxDataGrid'
-| 'dxTreeList'
-| 'dxPager'
-| 'dxRadioGroup'
-| 'dxScheduler'
-| 'dxTabs'
-| 'dxTagBox'
-| 'dxContextMenu'
-| 'dxDropDownMenu'
-| 'dxChart'
-| 'dxMenu'
-| 'dxPopup'
-| 'dxPopover'
-| 'dxSpeedDialAction'
-| 'dxSortable'
-| 'dxButton'
-| 'dxTextBox'
-| 'dxTextArea'
-| 'dxToolbar'
-| 'dxTreeView'
-| 'dxDateBox'
-| 'dxDateRangeBox'
-| 'dxLookup'
-| 'dxOverlay'
-| 'dxList'
-| 'dxHtmlEditor'
-| 'dxNumberBox'
-| 'dxValidator'
-| 'dxFileUploader'
-| 'dxSlider'
-| 'dxDropDownBox'
-| 'dxFileManager';
+import { WidgetName } from './widgetName';
 
 interface WidgetOptions {
   dxDataGrid: DataGridProperties;
