+++ packages/devextreme/testing/testcafe/helpers/disposeWidget.ts
@@ -1,0 +1,19 @@
+import { ClientFunction } from 'testcafe';
+import { WidgetName } from './widgetName';
+
+export default async function disposeWidget<TWidgetName extends WidgetName>(
+  componentName: TWidgetName,
+  selector = '#container',
+): Promise<void> {
+  await ClientFunction(
+    () => {
+      ($(`${selector}`) as any)[componentName]('dispose');
+    },
+    {
+      dependencies: {
+        componentName,
+        selector,
+      },
+    },
+  )();
+}
