+++ packages/devextreme/testing/testcafe/helpers/widgetName.ts
@@ -1,0 +1,52 @@
+export type WidgetName =
+'dxAccordion'
+| 'dxAutocomplete'
+| 'dxGallery'
+| 'dxButtonGroup'
+| 'dxCalendar'
+| 'dxCalendarView'
+| 'dxCheckBox'
+| 'dxColorBox'
+| 'dxDropDownButton'
+| 'dxDraggable'
+| 'dxTabPanel'
+| 'dxForm'
+| 'dxFilterBuilder'
+| 'dxSelectBox'
+| 'dxScrollable'
+| 'dxScrollView'
+| 'dxMultiView'
+| 'dxPivotGrid'
+| 'dxPivotGridFieldChooser'
+| 'dxDataGrid'
+| 'dxTreeList'
+| 'dxPager'
+| 'dxRadioGroup'
+| 'dxScheduler'
+| 'dxTabs'
+| 'dxTagBox'
+| 'dxContextMenu'
+| 'dxDropDownMenu'
+| 'dxChart'
+| 'dxMenu'
+| 'dxPopup'
+| 'dxPopover'
+| 'dxSpeedDialAction'
+| 'dxSortable'
+| 'dxButton'
+| 'dxTextBox'
+| 'dxTextArea'
+| 'dxToolbar'
+| 'dxTreeView'
+| 'dxDateBox'
+| 'dxDateRangeBox'
+| 'dxLookup'
+| 'dxOverlay'
+| 'dxList'
+| 'dxHtmlEditor'
+| 'dxNumberBox'
+| 'dxValidator'
+| 'dxFileUploader'
+| 'dxSlider'
+| 'dxDropDownBox'
+| 'dxFileManager';
