+++ packages/devextreme/testing/testcafe/tests/scheduler/appointmentForm/expressions.ts
@@ -1,5 +1,5 @@
-import { ClientFunction } from 'testcafe';
 import { createScreenshotsComparer } from 'devextreme-screenshot-comparer';
+import disposeWidget from '../../../helpers/disposeWidget';
 import createWidget from '../../../helpers/createWidget';
 import url from '../../../helpers/getPageUrl';
 import Scheduler from '../../../model/scheduler';
@@ -10,11 +10,6 @@ fixture.disablePageReloads`Appointment form: expressions`
 const SCHEDULER_SELECTOR = '#container';
 const TEST_TITLE = 'Test';
 const TEST_DESCRIPTION = 'Test description...';
-
-// TODO Vinogradov: Create a separate "disposeWidget" helper function.
-const disposeScheduler = async () => ClientFunction(() => {
-  ($(SCHEDULER_SELECTOR) as any).dxScheduler('dispose');
-}, { dependencies: { SCHEDULER_SELECTOR } })();
 
 // General test cases
 [
@@ -486,7 +481,9 @@ const disposeScheduler = async () => ClientFunction(() => {
         cellDuration: 240,
         ...options,
       });
-    }).after(async () => disposeScheduler());
+    }).after(async () => {
+      await disposeWidget('dxScheduler');
+    });
   });
 });
 
@@ -529,4 +526,6 @@ test(
     recurrenceEditMode: 'series',
     recurrenceRuleExpr: 'nestedA.nestedB.nestedC.recurrenceRuleCustom',
   });
-}).after(async () => disposeScheduler());
+}).after(async () => {
+  await disposeWidget('dxScheduler');
+});
