+++ packages/devextreme/testing/testcafe/tests/scheduler/viewOffset/currentTimeIndicator.ts
@@ -1,5 +1,5 @@
 import { createScreenshotsComparer } from 'devextreme-screenshot-comparer';
-import { ClientFunction } from 'testcafe';
+import disposeWidget from '../../../helpers/disposeWidget';
 import createWidget from '../../../helpers/createWidget';
 import url from '../../../helpers/getPageUrl';
 import Scheduler from '../../../model/scheduler';
@@ -112,9 +112,6 @@ endDayHour: ${endDayHour}
       endDayHour,
     });
   }).after(async () => {
-    // TODO Vinogradov: Create a separate "disposeWidget" helper function.
-    await ClientFunction(() => {
-      ($(SCHEDULER_SELECTOR) as any).dxScheduler('dispose');
-    }, { dependencies: { SCHEDULER_SELECTOR } })();
+    await disposeWidget('dxScheduler');
   });
 });
